requirements.txt
Flask==3.1.0
Flask-Cors==4.0.1
python-dotenv==0.21.0
SQLAlchemy==1.4.47
psycopg2-binary==2.9.6
gunicorn==22.0.0
